+++ jss/rprobitb_oelschlaeger_bauer.R
@@ -4,163 +4,153 @@
 ### code chunk number 1: preliminaries
 ###################################################
 options(prompt = "> ", continue = "+  ", width = 70, useFancyQuotes = FALSE)
-library(RprobitB)
+#library(RprobitB) # UNCOMMENT!
 
 
 ###################################################
-### code chunk number 2: prepare-data-call (eval = FALSE)
+### code chunk number 2: prepare data call (eval = FALSE)
 ###################################################
 ## data <- prepare_data(form = form, choice_data = choice_data)
 
 
 ###################################################
-### code chunk number 3: overview-Train-data
-###################################################
-data("Train", package = "mlogit")
-str(Train)
-
-
-###################################################
-### code chunk number 4: Train-formula
-###################################################
-form <- choice ~ price + time + comfort + change | 0
-
-
-###################################################
-### code chunk number 5: Train-re
-###################################################
-re <- c("price","time")
-
-
-###################################################
-### code chunk number 6: check-form
-###################################################
-check_form(form = form, re = re)
-
-
-###################################################
-### code chunk number 7: prepare-Train-data
-###################################################
-data <- prepare_data(form = form, choice_data = Train, re = re, id = "id", idc = "choiceid")
-
-
-###################################################
-### code chunk number 8: summary-plot-Train-data
-###################################################
-summary(data)
-plot(data)
-
-
-###################################################
-### code chunk number 9: simulate-choices-call (eval = FALSE)
-###################################################
-## data <- simulate_choices(form = form, N = N, T = T, J = J)
-
-
-###################################################
-### code chunk number 10: data-sim-meta
-###################################################
-N <- 100
-T <- 10
-alternatives <- c("A", "B")
-form <- choice ~ var1 | var2 | var3
-re <- c("ASC", "var2")
-
-
-###################################################
-### code chunk number 11: data-sim-overview
-###################################################
-overview_effects(form = form, re = re, alternatives = alternatives)
-
-
-###################################################
-### code chunk number 12: data-sim
-###################################################
-data <- simulate_choices(
-  form = form,
-  N = N,
-  T = T,
-  J = 2,
-  re = re,
-  alternatives = alternatives,
-  seed = 1,
-  alpha = c(-1,0,1),
-  b = matrix(c(2,-0.5), ncol = 1)
-)
-summary(data)
-
-
-###################################################
-### code chunk number 13: data-sim-plot-by-choice
-###################################################
-plot(data, by_choice = TRUE)
-
-
-###################################################
-### code chunk number 14: data-split-deciders
-###################################################
-train_test(data, test_proportion = 0.3, by = "N")
-
-
-###################################################
-### code chunk number 15: data-split-occasions
-###################################################
-train_test(data, test_number = 2, by = "T", random = TRUE, seed = 1)
-
-
-###################################################
-### code chunk number 16: mcmc-call (eval = FALSE)
-###################################################
-## mcmc(data = data)
-
-
-###################################################
-### code chunk number 17: transform-Train
+### code chunk number 3: example 1 train overview
 ###################################################
 data("Train", package = "mlogit")
 Train$price_A <- Train$price_A / 100 * 2.20371
 Train$price_B <- Train$price_B / 100 * 2.20371
 Train$time_A <- Train$time_A / 60
 Train$time_B <- Train$time_B / 60
+str(Train)
 
 
 ###################################################
-### code chunk number 18: Train-fit (eval = FALSE)
+### code chunk number 4: example 1 train formula
 ###################################################
-## form <- choice ~ price + time + change + comfort | 0
-## data <- prepare_data(form = form, choice_data = Train)
-## model_train <- mcmc(
-##   data = data,
-##   scale = list("parameter" = "a", index = 1, value = -1)
-## )
+form <- choice ~ price + time + comfort + change | 0
 
 
 ###################################################
-### code chunk number 19: load-model-Train
+### code chunk number 5: example 1 train prepare
+###################################################
+data_train <- prepare_data(
+  form = form,
+  choice_data = Train,
+  id = "id",
+  idc = "choiceid"
+  )
+summary(data_train)
+
+
+###################################################
+### code chunk number 6: train-data
+###################################################
+plot(data_train)
+
+
+###################################################
+### code chunk number 7: simulate choices call (eval = FALSE)
+###################################################
+## data <- simulate_choices(form = form, N = N, T = T, J = J)
+
+
+###################################################
+### code chunk number 8: example 2 sim meta
+###################################################
+N <- 100
+T <- 10
+alternatives <- c("alt1", "alt2")
+base_alternative <- "alt2"
+form <- choice ~ var1 | var2 | var3
+re <- c("ASC","var2")
+
+
+###################################################
+### code chunk number 9: example 2 sim overview effects
+###################################################
+overview_effects(
+  form = form,
+  re = re,
+  alternatives = alternatives,
+  base_alternative = base_alternative
+)
+
+
+###################################################
+### code chunk number 10: example 2 sim simulation
+###################################################
+data_sim <- simulate_choices(
+  form = form,
+  N = N,
+  T = T,
+  J = 2,
+  re = re,
+  alternatives = alternatives,
+  base_alternative = base_alternative,
+  seed = 1,
+  alpha = c(-1,0,1),
+  C = 2,
+  s = c(0.7,0.3),
+  b = matrix(c(2,-0.5,1,1), ncol = 2)
+)
+
+
+###################################################
+### code chunk number 11: sim-data
+###################################################
+plot(data_sim, by_choice = TRUE)
+
+
+###################################################
+### code chunk number 12: data-split-deciders
+###################################################
+train_test(data_sim, test_proportion = 0.3, by = "N")
+
+
+###################################################
+### code chunk number 13: data-split-occasions
+###################################################
+train_test(data_sim, test_number = 2, by = "T", random = TRUE, seed = 1)
+
+
+###################################################
+### code chunk number 14: fit_model call (eval = FALSE)
+###################################################
+## fit_model(data = data)
+
+
+###################################################
+### code chunk number 15: example 1 train fit (eval = FALSE)
+###################################################
+## model_train <- fit_model(data = data_train, scale = price ~ -1)
+
+
+###################################################
+### code chunk number 16: example 1 train load model
 ###################################################
 data(model_train, package = "RprobitB")
 
 
 ###################################################
-### code chunk number 20: coef-model-Train
+### code chunk number 17: coef-model-Train
 ###################################################
 coef(model_train)
 
 
 ###################################################
-### code chunk number 21: plot-coef-model-train
+### code chunk number 18: plot-coef-model-train
 ###################################################
 plot(coef(model_train), sd = 3)
 
 
 ###################################################
-### code chunk number 22: str-gibbs-samples
+### code chunk number 19: example 1 train gibbs samples
 ###################################################
 str(model_train$gibbs_samples, max.level = 2, give.attr = FALSE)
 
 
 ###################################################
-### code chunk number 23: summary-model-train
+### code chunk number 20: example 1 train summary
 ###################################################
 summary(model_train,
         FUN = c("mean"        = mean,
@@ -172,64 +162,64 @@ summary(model_train,
 
 
 ###################################################
-### code chunk number 24: plot-trace-model-train
+### code chunk number 21: model-train-trace
 ###################################################
-par(mfrow = c(2,1))
+par(mfrow = c(1,2))
 plot(model_train, type = "trace")
 
 
 ###################################################
-### code chunk number 25: plot-acf-model-train
+### code chunk number 22: model-train-acf
 ###################################################
 par(mfrow = c(2,3))
 plot(model_train, type = "acf")
 
 
 ###################################################
-### code chunk number 26: transform-model-train-B
+### code chunk number 23: example 1 train transform B
 ###################################################
 model_train <- transform(model_train, B = 1)
 
 
 ###################################################
-### code chunk number 27: transform-model-train-Q
+### code chunk number 24: example 1 train transform Q
 ###################################################
 model_train <- transform(model_train, Q = 100)
 
 
 ###################################################
-### code chunk number 28: transform-model-train-scale
+### code chunk number 25: example 1 train transform scale
 ###################################################
-model_train <- transform(model_train, scale = list(parameter = "s", index = 1, value = 1))
+model_train <- transform(model_train, scale = Sigma_1 ~ 1)
 
 
 ###################################################
-### code chunk number 29: estimate-elec-model (eval = FALSE)
+### code chunk number 26: example 3 elec estimation (eval = FALSE)
 ###################################################
 ## data("Electricity", package = "mlogit")
 ## Electricity <- as_cov_names(Electricity, c("pf","cl","loc","wk","tod","seas"), 1:4)
-## data <- prepare_data(
+## data_elec <- prepare_data(
 ##   form = choice ~ pf + cl + loc + wk + tod + seas | 0,
 ##   choice_data = Electricity,
 ##   re = c("cl","loc","wk","tod","seas")
 ##   )
-## model_elec <- mcmc(data, R = 1000, scale = list(parameter = "a", index = 1, value = -1))
+## model_elec <- fit_model(data_elec, R = 5000, scale = pf ~ -1)
 
 
 ###################################################
-### code chunk number 30: load-model-ele
+### code chunk number 27: example 3 elec load model
 ###################################################
 data(model_elec, package = "RprobitB")
 
 
 ###################################################
-### code chunk number 31: coef-model-elec
+### code chunk number 28: coef-model-elec
 ###################################################
 coef(model_elec)
 
 
 ###################################################
-### code chunk number 32: compute-mixdistr-share
+### code chunk number 29: example 3 elec share mixing distribution
 ###################################################
 cl_mu <- coef(model_elec)["cl","mean"]
 cl_sd <- sqrt(coef(model_elec)["cl","var"])
@@ -237,19 +227,26 @@ pnorm(cl_mu / cl_sd)
 
 
 ###################################################
-### code chunk number 33: cov-mixdistr
+### code chunk number 30: example 3 elec correlation mixing distribution
 ###################################################
 cov_mix(model_elec, cor = TRUE)
 
 
 ###################################################
-### code chunk number 34: plot-mixture-model-elec
+### code chunk number 31: example 2 sim model fit
 ###################################################
-plot(model_elec, type = "mixture")
+model_sim <- fit_model(data_sim, R = 1000, latent_classes = list("C" = 2))
+summary(model_sim)
 
 
 ###################################################
-### code chunk number 35: sim-dirichlet
+### code chunk number 32: model-sim-mixture
+###################################################
+plot(model_sim, type = "mixture")
+
+
+###################################################
+### code chunk number 33: sim-dirichlet
 ###################################################
 set.seed(1)
 P_r <- 2
@@ -265,7 +262,7 @@ z_true <- rep(1:3, times = N)
 
 
 ###################################################
-### code chunk number 36: dirichlet-prior
+### code chunk number 34: dirichlet-prior
 ###################################################
 delta <- 0.1
 xi <- numeric(P_r)
@@ -275,7 +272,7 @@ Theta <- diag(P_r)
 
 
 ###################################################
-### code chunk number 37: dirichlet-inits
+### code chunk number 35: dirichlet-inits
 ###################################################
 z <- rep(1, ncol(beta))
 C <- 1
@@ -284,7 +281,7 @@ Omega <- matrix(rep(diag(P_r), C), nrow = P_r*P_r, ncol = C)
 
 
 ###################################################
-### code chunk number 38: dirichlet-process-app
+### code chunk number 36: dirichlet-process-app
 ###################################################
 for(r in 1:100){
   dp <- RprobitB:::update_classes_dp(
@@ -297,181 +294,10 @@ for(r in 1:100){
 
 
 ###################################################
-### code chunk number 39: dirichlet-example-plot
+### code chunk number 37: dirichlet-example-plot
 ###################################################
 par(mfrow = c(1,2))
 plot(t(beta), xlab = bquote(beta[1]), ylab = bquote(beta[2]), pch = 19)
 RprobitB:::plot_class_allocation(beta, z, b, Omega, r = 100, perc = 0.95)
-
-
-###################################################
-### code chunk number 40: load-model-train
-###################################################
-data("model_train", package = "RprobitB")
-
-
-###################################################
-### code chunk number 41: predict-model-train
-###################################################
-predict(model_train)
-
-
-###################################################
-### code chunk number 42: predict-model-train-indlevel
-###################################################
-pred <- predict(model_train, overview = FALSE)
-head(pred, n = 10)
-
-
-###################################################
-### code chunk number 43: model-train-covs
-###################################################
-get_cov(model_train, id = 1, idc = 8)
-
-
-###################################################
-### code chunk number 44: model-train-coeffs
-###################################################
-coef(model_train)
-
-
-###################################################
-### code chunk number 45: model-train-Sigma
-###################################################
-point_estimates(model_train)$Sigma
-
-
-###################################################
-### code chunk number 46: roc-example
-###################################################
-library(plotROC)
-ggplot(data = pred, aes(m = A, d = ifelse(true == "A", 1, 0))) +
-  geom_roc(n.cuts = 20, labels = FALSE) +
-  style_roc(theme = theme_grey)
-
-
-###################################################
-### code chunk number 47: predict-model-train-given-covs-1
-###################################################
-predict(
-  model_train,
-  data = data.frame("price_A" = c(100,110),
-                    "price_B" = c(100,100)),
-  overview = FALSE)
-
-
-###################################################
-### code chunk number 48: predict-model-train-given-covs-2
-###################################################
-predict(
-  model_train,
-  data = data.frame("price_A"   = c(100,110),
-                    "comfort_A" = c(1,0),
-                    "price_B"   = c(100,100),
-                    "comfort_B" = c(1,1)),
-  overview = FALSE)
-
-
-###################################################
-### code chunk number 49: load-model-train
-###################################################
-data("model_train", package = "RprobitB")
-model_train
-
-
-###################################################
-### code chunk number 50: nested-model-train (eval = FALSE)
-###################################################
-## model_train_sparse <- nested_model(model_train, form = choice ~ price | 0)
-
-
-###################################################
-### code chunk number 51: load-train-sparse-model
-###################################################
-data("model_train_sparse", package = "RprobitB")
-
-
-###################################################
-### code chunk number 52: model-selection-example
-###################################################
-model_selection(model_train, model_train_sparse,
-                criteria = c("npar", "LL", "AIC", "BIC", "WAIC", "MMLL", "BF", "pred_acc"))
-
-
-###################################################
-### code chunk number 53: npar-example
-###################################################
-npar(model_train, model_train_sparse)
-
-
-###################################################
-### code chunk number 54: oglik-example
-###################################################
-logLik(model_train)
-logLik(model_train_sparse)
-
-
-###################################################
-### code chunk number 55: aic-example
-###################################################
-AIC(model_train, model_train_sparse, k = 2)
-
-
-###################################################
-### code chunk number 56: bic-example
-###################################################
-BIC(model_train, model_train_sparse)
-
-
-###################################################
-### code chunk number 57: compute-psi-train (eval = FALSE)
-###################################################
-## model_train <- compute_p_si(model_train)
-
-
-###################################################
-### code chunk number 58: waic-example
-###################################################
-WAIC(model_train)
-WAIC(model_train_sparse)
-
-
-###################################################
-### code chunk number 59: waictrace
-###################################################
-plot(WAIC(model_train))
-plot(WAIC(model_train_sparse))
-
-
-###################################################
-### code chunk number 60: mml_train
-###################################################
-model_train <- mml(model_train)
-model_train$mml
-attr(model_train$mml, "mmll")
-
-
-###################################################
-### code chunk number 61: mmltrace
-###################################################
-plot(model_train$mml, log = TRUE)
-
-
-###################################################
-### code chunk number 62: arithmetic_mean_estimator (eval = FALSE)
-###################################################
-## model_train <- mml(model_train, S = 1000, recompute = TRUE)
-
-
-###################################################
-### code chunk number 63: bayes-factor-example
-###################################################
-model_selection(model_train, model_train_sparse, criteria = c("BF"))
-
-
-###################################################
-### code chunk number 64: pred-acc-example
-###################################################
-pred_acc(model_train, model_train_sparse)